refactor: reorganized the socket code on the client into its own module
