+++ frontend/src/components/generative/ActorContent.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useMemo, memo, useCallback } from "react";
 import type { BaseMessage } from "@/store/useMessageStore";
-import type { Actor } from "@/types/envelopeType";
+import type { Actor } from "@/socket/envelopeType";
 import { actorRegistry } from "./Registry";
 import { TabsContent } from "../ui/tabs";
